Shared GameTiming no longer depends on INetManager Adds ClientGameTiming with prediction and local/server time methods.
